+++ src/cli.h
@@ -25,19 +25,19 @@ typedef volatile int vint;
 
 extern volatile int* cli_timer_flag;
 extern volatile int cli__reset;
-void cli_progress_add(SEXP bar, int inc);
-SEXP cli_progress_bar(vint **ptr, int total, SEXP config);
+void cli_progress_add(SEXP bar, double inc);
+SEXP cli_progress_bar(vint **ptr, double total, SEXP config);
 void cli_progress_done(SEXP bar);
 void cli_progress_init_timer(vint **ptr);
 int cli_progress_num();
-void cli_progress_set(SEXP bar, int set);
+void cli_progress_set(SEXP bar, double set);
 void cli_progress_set_clear(SEXP bar, int);
 void cli_progress_set_format(SEXP bar, const char *name);
 void cli_progress_set_name(SEXP bar, const char *name);
 void cli_progress_set_status(SEXP bar, const char *name);
 void cli_progress_set_type(SEXP bar, const char *name);
 void cli_progress_sleep(int s, long ns);
-void cli_progress_update(SEXP bar, int set, int inc, int force);
+void cli_progress_update(SEXP bar, double set, double inc, int force);
 
 SEXP cli__progress_update(SEXP bar);
 SEXP clic_progress_along(SEXP seq, SEXP bar);
